# worker/openapi.worker.yaml
openapi: 3.0.3
info:
  title: Camofleet Worker API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /sessions:
    get:
      summary: List sessions
      responses:
        '200':
          description: Array of active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionDetail'
    post:
      summary: Create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
  /sessions/{id}:
    get:
      summary: Get session detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '404':
          description: Session not found
    delete:
      summary: Delete session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session terminated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDeleteResponse'
        '404':
          description: Session not found
  /sessions/{id}/touch:
    post:
      summary: Extend TTL
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
        version:
          type: string
        checks:
          type: object
          additionalProperties:
            type: string
    SessionCreateRequest:
      type: object
      properties:
        browser:
          type: string
          enum: [chromium, firefox, webkit]
        headless:
          type: boolean
        idle_ttl_seconds:
          type: integer
          minimum: 30
          maximum: 3600
        labels:
          type: object
          additionalProperties:
            type: string
      additionalProperties: false
    SessionDetail:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        browser:
          type: string
        headless:
          type: boolean
        idle_ttl_seconds:
          type: integer
        labels:
          type: object
          additionalProperties:
            type: string
        worker_id:
          type: string
        ws_endpoint:
          type: string
        vnc:
          type: object
          properties:
            ws:
              type: string
              nullable: true
            http:
              type: string
              nullable: true
            password_protected:
              type: boolean
      required: [id, status, created_at, last_seen_at, browser, headless, idle_ttl_seconds, labels, worker_id, ws_endpoint, vnc]
    SessionDeleteResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
      required: [id, status]